+++ packages/atom/src/index.ts
@@ -2,4 +2,5 @@ export * from "./api";
 export * from "./atom";
 export * from "./cursor";
 export * from "./history";
+export * from "./transacted";
 export * from "./view";