+++ packages/matrices/src/m33-m44.ts
@@ -1,17 +1,34 @@
-import { setS3, setS4, ZERO4 } from "@thi.ng/vectors";
+import { setC } from "@thi.ng/vectors";
 import { MatOpM } from "./api";
 
 /**
- * Converts M33 to M44 and writes result to `out`.
+ * Converts 3x3 to 4x4 matrix and writes result to `out`. Creates new
+ * matrix if `out` is `null`.
  *
  * @param out
  * @param m33
  */
-export const mat33to44: MatOpM = (out, m33) => (
-    !out && (out = []),
-    setS3(out, m33, 0, 0),
-    setS3(out, m33, 4, 3),
-    setS3(out, m33, 8, 6),
-    setS3(out, ZERO4, 12),
-    setS4(out, [0, 0, 0, 1], 3, 0, 4)
-);
+export const mat33to44: MatOpM = (out, m33) =>
+    setC(
+        out || [],
+        // x
+        m33[0],
+        m33[1],
+        m33[2],
+        0,
+        // y
+        m33[3],
+        m33[4],
+        m33[5],
+        0,
+        // z
+        m33[6],
+        m33[7],
+        m33[8],
+        0,
+        // w
+        0,
+        0,
+        0,
+        1
+    );