+++ packages/matrices/src/translation.ts
@@ -2,7 +2,7 @@ import { ReadonlyVec, setC, setC6 } from "@thi.ng/vectors";
 import { Mat } from "./api";
 
 /**
- * Constructs a M23 translation matrix.
+ * Constructs a 2x3 translation matrix.
  *
  * @param out
  * @param v
@@ -11,7 +11,7 @@ export const translation23 = (m: Mat | null, v: ReadonlyVec) =>
     setC6(m || [], 1, 0, 0, 1, v[0], v[1]);
 
 /**
- * Constructs a M44 translation matrix.
+ * Constructs a 4x4 translation matrix.
  *
  * @param out
  * @param v