+++ packages/webgl/src/index.ts
@@ -1,14 +1,24 @@
-export * from "./api";
+export * from "./api/blend";
+export * from "./api/buffers";
+export * from "./api/canvas";
+export * from "./api/ext";
+export * from "./api/glsl";
+export * from "./api/material";
+export * from "./api/model";
+export * from "./api/shader";
+export * from "./api/texture";
+
 export * from "./buffer";
 export * from "./canvas";
+export * from "./checks";
 export * from "./draw";
 export * from "./error";
 export * from "./fbo";
-export * from "./gpgpu";
 export * from "./material";
 export * from "./matrices";
-export * from "./pipeline";
+export * from "./multipass";
 export * from "./rbo";
+export * from "./readpixels";
 export * from "./shader";
 export * from "./syntax";
 export * from "./texture";
@@ -16,6 +26,7 @@ export * from "./utils";
 
 export * from "./shaders/lambert";
 export * from "./shaders/phong";
+export * from "./shaders/pipeline";
 
 export * from "./textures/checkerboard";
 export * from "./textures/stripes";