+++ packages/pointfree-lang/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/thi-ng/umbrella/compare/@thi.ng/pointfree-lang@1.1.4...@thi.ng/pointfree-lang@1.1.5) (2019-09-21)
+
+
+### Bug Fixes
+
+* **pointfree-lang:** update imports ([8de1366](https://github.com/thi-ng/umbrella/commit/8de1366))
+
+
+
+
+
 ## [1.1.4](https://github.com/thi-ng/umbrella/compare/@thi.ng/pointfree-lang@1.1.3...@thi.ng/pointfree-lang@1.1.4) (2019-08-21)
 
 **Note:** Version bump only for package @thi.ng/pointfree-lang