+++ packages/transducers-binary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.5](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-binary@0.4.4...@thi.ng/transducers-binary@0.4.5) (2019-09-21)
+
+**Note:** Version bump only for package @thi.ng/transducers-binary
+
+
+
+
+
 ## [0.4.4](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-binary@0.4.3...@thi.ng/transducers-binary@0.4.4) (2019-08-21)
 
 **Note:** Version bump only for package @thi.ng/transducers-binary